Convert raw data into a tech exposure data frame

The as_tech_exposure_data function converts raw data into a tech exposure data frame

as_tech_exposure_data(
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  investor_name,
  portfolio_name,
  start_year,
  green_techs = c("RenewablesCap", "HydroCap", "NuclearCap", "Hybrid", "Electric",
    "FuelCell", "Hybrid_HDV", "Electric_HDV", "FuelCell_HDV", "Ac-Electric Arc Furnace",
    "Dc-Electric Arc Furnace"),
  select_scenario,
  select_scenario_auto,
  select_scenario_shipping,
  select_scenario_other,
  all_tech_levels,
  equity_market_levels,
  dataframe_translations,
  language_select = "EN"
)

Arguments

equity_results_portfolio

equity_results_portfolio

bonds_results_portfolio

bonds_results_portfolio

indices_equity_results_portfolio

indices_equity_results_portfolio

indices_bonds_results_portfolio

indices_bonds_results_portfolio

investor_name

investor_name

portfolio_name

portfolio_name

start_year

start_year

green_techs

green_techs

select_scenario

select_scenario

select_scenario_auto

select_scenario_auto

select_scenario_shipping

select_scenario_shipping

select_scenario_other

select_scenario_other

all_tech_levels

all_tech_levels

equity_market_levels

equity_market_levels

dataframe_translations

dataframe_translations

language_select

two letter code for language (single string; default = "EN")