Convert raw data into a trajectory alignment data frame

The as_trajectory_alignment_data function converts raw data into a trajectory alignment data frame

Usage

as_trajectory_alignment_data(
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  investor_name,
  portfolio_name,
  tech_roadmap_sectors,
  scen_geo_levels,
  all_tech_levels,
  year_horizon = 5,
  dataframe_translations,
  language_select = "en"
)

Arguments

equity_results_portfolio

equity_results_portfolio

bonds_results_portfolio

bonds_results_portfolio

indices_equity_results_portfolio

indices_equity_results_portfolio

indices_bonds_results_portfolio

indices_bonds_results_portfolio

investor_name

investor_name

portfolio_name

portfolio_name

tech_roadmap_sectors

tech_roadmap_sectors

scen_geo_levels

scen_geo_levels

all_tech_levels

all_tech_levels

year_horizon

number of years to include from the start year (single integer/numeric)

dataframe_translations

dataframe_translations

language_select

two letter code for language (single string; default = "EN")